# Optimize MoE topk with torch compile